% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_table_gt_functions.R
\name{add_color_to_lift}
\alias{add_color_to_lift}
\title{Adding color to the lift metric}
\usage{
add_color_to_lift(performance_dat, metric, color)
}
\arguments{
\item{performance_dat}{the original performance data}

\item{metric}{the required metric}

\item{color}{the required color}
}
\description{
Adding color to the lift metric
}
\keyword{internal}
