% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_table_gt_functions.R
\name{add_color_to_net_benifit}
\alias{add_color_to_net_benifit}
\title{Adding Color to Net Benifit metric}
\usage{
add_color_to_net_benifit(performance_data)
}
\arguments{
\item{performance_data}{an rtichoke Performance Data}
}
\value{

}
\description{
Adding Color to Net Benifit metric
}
\keyword{internal}
