% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_table_gt_functions.R
\name{add_colors_to_performance_dat}
\alias{add_colors_to_performance_dat}
\title{Adding color to Performance Data}
\usage{
add_colors_to_performance_dat(performance_dat)
}
\arguments{
\item{performance_dat}{the original performance data}
}
\value{

}
\description{
Adding color to Performance Data
}
\keyword{internal}
