% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_for_performance_metrics.R
\name{add_decision_curve_reference_lines}
\alias{add_decision_curve_reference_lines}
\title{Add reference lines to Precision Recall Curve}
\usage{
add_decision_curve_reference_lines(decision_curve)
}
\arguments{
\item{decision_curve}{a ggplot object of precision recall curve}
}
\description{
Add reference lines to Precision Recall Curve
}
