% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_for_performance_metrics.R
\name{add_gains_curve_reference_lines}
\alias{add_gains_curve_reference_lines}
\title{Add reference lines to Gains Curve}
\usage{
add_gains_curve_reference_lines(gains_curve, prevalence)
}
\arguments{
\item{gains_curve}{a ggplot object of a Gains Curve}

\item{prevalence}{the prevalence of the outcome}
}
\description{
Add reference lines to Gains Curve
}
