% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_table.R
\name{add_html_color_to_model_for_subtitle}
\alias{add_html_color_to_model_for_subtitle}
\title{Add html color to model for subtitle}
\usage{
add_html_color_to_model_for_subtitle(model_name, model_color)
}
\arguments{
\item{model_name}{the name of the model}

\item{model_color}{the color of the model}
}
\description{
Add html color to model for subtitle
}
\keyword{internal}
