% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_for_performance_metrics.R
\name{add_interactive_marker_to_plotly}
\alias{add_interactive_marker_to_plotly}
\title{Add Interactive Marker to Plotly}
\usage{
add_interactive_marker_to_plotly(
  plot_tichoke_ly_ob,
  multiple_models = F,
  main_slider = "threshold"
)
}
\arguments{
\item{plot_tichoke_ly_ob}{}

\item{multiple_models}{}

\item{main_slider}{}
}
\description{
Add Interactive Marker to Plotly
}
