% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_for_performance_metrics.R
\name{add_lift_curve_reference_lines}
\alias{add_lift_curve_reference_lines}
\title{Add reference lines to ROC CURVE}
\usage{
add_lift_curve_reference_lines(lift_curve)
}
\arguments{
\item{lift_curve}{a ggplot object of lift curve}
}
\description{
Add reference lines to ROC CURVE
}
