% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_for_performance_metrics.R
\name{add_lines_to_plotly}
\alias{add_lines_to_plotly}
\title{Add lines to Plotly}
\usage{
add_lines_to_plotly(plot_tichoke_ly_ob, multiple_models = F)
}
\arguments{
\item{plot_tichoke_ly_ob}{}

\item{multiple_models}{}
}
\description{
Add lines to Plotly
}
