% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_helper_functions.R
\name{add_markers_and_lines_to_plotly}
\alias{add_markers_and_lines_to_plotly}
\title{Adding markers and lines to plotly}
\usage{
add_markers_and_lines_to_plotly(plotly_object, performance_data_type)
}
\arguments{
\item{performance_data_type}{}
}
\description{
Adding markers and lines to plotly
}
\keyword{internal}
