% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_for_performance_metrics.R
\name{add_precision_recall_curve_reference_lines}
\alias{add_precision_recall_curve_reference_lines}
\title{Add reference lines to Precision Recall Curve}
\usage{
add_precision_recall_curve_reference_lines(precision_recall_curve, prevalence)
}
\arguments{
\item{precision_recall_curve}{a ggplot object of precision recall curve}

\item{prevalence}{the prevalence of the outcome}
}
\description{
Add reference lines to Precision Recall Curve
}
