% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_for_performance_metrics.R
\name{add_prevalence_layers_to_gains_curve}
\alias{add_prevalence_layers_to_gains_curve}
\title{Title}
\usage{
add_prevalence_layers_to_gains_curve(prevalence, col_value)
}
\arguments{
\item{prevalence}{}

\item{col_value}{}
}
\description{
Title
}
