% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_reference_lines_to_ggplot}
\alias{add_reference_lines_to_ggplot}
\title{Add reference lines to ggplot curve}
\usage{
add_reference_lines_to_ggplot(ggplot_curve, reference_lines)
}
\arguments{
\item{ggplot_curve}{a non interactive ggplot curve}

\item{reference_lines}{dataframe of reference lines}
}
\description{
Add reference lines to ggplot curve
}
\keyword{internal}
