% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_for_performance_metrics.R
\name{add_roc_curve_reference_lines}
\alias{add_roc_curve_reference_lines}
\title{Add reference lines to ROC CURVE}
\usage{
add_roc_curve_reference_lines(roc_curve)
}
\arguments{
\item{roc_curve}{a ggplot object of roc curve}
}
\description{
Add reference lines to ROC CURVE
}
