% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_table.R
\name{add_zebra_colors_to_gt_table}
\alias{add_zebra_colors_to_gt_table}
\title{Add Zebra colors to gt table}
\usage{
add_zebra_colors_to_gt_table(performance_table, add_zebra_colors)
}
\arguments{
\item{performance_table}{gt performance table}

\item{add_zebra_colors}{add zebra colors or keep table as it is (FALSE)}
}
\description{
Add Zebra colors to gt table
}
\keyword{internal}
