% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{arrange_estimated_probabilities_to_long_format}
\alias{arrange_estimated_probabilities_to_long_format}
\title{Arrange estimated probabilities to long format}
\usage{
arrange_estimated_probabilities_to_long_format(probs)
}
\arguments{
\item{probs}{a list of vectors of estimated probabilities
(one for each model or one for each population)}
}
\description{
Arrange estimated probabilities to long format
}
\examples{
\dontrun{
arrange_estimated_probabilities_to_long_format(
  probs = list("First Model" = example_dat$estimated_probabilities)
)
}
}
\keyword{internal}
