% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_table_gt_functions.R
\name{bar_chart}
\alias{bar_chart}
\title{Creating html bar}
\usage{
bar_chart(value, display, color = "red", digits = 0, no_round = FALSE)
}
\arguments{
\item{value}{original value}

\item{display}{displayed value}

\item{color}{color of the bar}

\item{digits}{number of digits}

\item{no_round}{no rounding}
}
\description{
Creating html bar
}
\keyword{internal}
