% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs.R
\name{check_chosen_threshold_input}
\alias{check_chosen_threshold_input}
\title{Check chosen threshold input}
\usage{
check_chosen_threshold_input(chosen_threshold)
}
\description{
Check chosen threshold input
}
\examples{
\dontrun{
check_chosen_threshold_input(0.23)
check_chosen_threshold_input(1.02)

} 
}
\keyword{internal}
