% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs.R
\name{check_curve_input}
\alias{check_curve_input}
\title{Check curve input}
\usage{
check_curve_input(curve)
}
\arguments{
\item{curve}{discrimination curve}
}
\description{
Check curve input
}
\examples{
\dontrun{
check_curve_input("roc")
check_curve_input("lift")
}
}
\keyword{internal}
