% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs.R
\name{check_performance_data_stratification}
\alias{check_performance_data_stratification}
\title{Check cheson threshold input}
\usage{
check_performance_data_stratification(performance_data)
}
\description{
Check cheson threshold input
}
\examples{
\dontrun{
check_performance_data_stratification(one_pop_one_model_as_a_vector)
check_performance_data_stratification(
  one_pop_one_model_as_a_vector_enforced_percentiles_symmetry
)
}
}
\keyword{internal}
