% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_helper_functions.R
\name{check_performance_data_type_for_plotly}
\alias{check_performance_data_type_for_plotly}
\title{Check Performance type for plotly}
\usage{
check_performance_data_type_for_plotly(performance_data)
}
\arguments{
\item{performance_data}{an rtichoke Performance Data}
}
\description{
Check Performance type for plotly
}
\keyword{internal}
