% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs.R
\name{check_stratified_by_input}
\alias{check_stratified_by_input}
\title{Check stratified_by input}
\usage{
check_stratified_by_input(stratified_by)
}
\arguments{
\item{stratified_by}{stratified by probability threshold or ppcr}
}
\description{
Check stratified_by input
}
\examples{
\dontrun{
check_curve_input("probability_threshold")
check_curve_input("ppcr")
}
}
\keyword{internal}
