% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_helper_functions.R
\name{create_color_reference_lines_vector}
\alias{create_color_reference_lines_vector}
\title{Creating color reference lines vector}
\usage{
create_color_reference_lines_vector(color_populations_vector, curve)
}
\arguments{
\item{color_populations_vector}{color population vector}

\item{curve}{a curve}
}
\value{

}
\description{
Creating color reference lines vector
}
\keyword{internal}
