% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{create_dat_for_smooth_calibration}
\alias{create_dat_for_smooth_calibration}
\title{Creating dat for smooth calibration}
\usage{
create_dat_for_smooth_calibration(probs, real, deciles_dat)
}
\arguments{
\item{probs}{a vector of estimated probabilities or a list of vectors of that
kind (one for each model)}

\item{real}{a vector of binary outcomes or a list of vectors of that
kind (one for each population)}

\item{deciles_dat}{data of deciles}
}
\description{
Arrange estimated probabilities to long format
}
\examples{
\dontrun{
deciles_dat <- purrr::map_df(list("Model 1" = example_dat$estimated_probabilities),
                             ~ make_deciles_dat(.x, example_dat$outcome),
                             .id = "model"
) \%>\% 
  mutate(model  = forcats::fct_inorder(factor(model)))

create_dat_for_smooth_calibration(
  list("Model 1" = example_dat$estimated_probabilities),
  real = example_dat$outcome,
  deciles_dat
)
}
}
\keyword{internal}
