% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_for_performance_metrics.R
\name{create_ggplot_for_performance_metrics}
\alias{create_ggplot_for_performance_metrics}
\title{Create ggplot for performance metrics}
\usage{
create_ggplot_for_performance_metrics(
  performance_table,
  x_perf_metric,
  y_perf_metric,
  col_values = c("#E69F00", "#56B4E9", "#F0E442", "#0072B2", "#CC79A7")
)
}
\arguments{
\item{performance_table}{an rtichoke performance table}

\item{x_perf_metric}{a performance metrice for the x axis}

\item{y_perf_metric}{a performance metrice for the y axis}

\item{col_values}{color palette}
}
\description{
Makes a ggplot for the metrices
}
