% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_helper_functions.R
\name{create_linetype_reference_vector}
\alias{create_linetype_reference_vector}
\title{Creating linetype reference lines vector}
\usage{
create_linetype_reference_vector(color_populations_vector, curve)
}
\arguments{
\item{color_populations_vector}{color population vector}

\item{curve}{a curve}
}
\value{

}
\description{
Creating linetype reference lines vector
}
\keyword{internal}
