% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_performance_table.R
\name{create_performance_table_multiple_pop}
\alias{create_performance_table_multiple_pop}
\title{Create Performance Table for Multiple Populations}
\usage{
create_performance_table_multiple_pop(
  list_of_lists_of_different_pop,
  by = 0.01,
  enforce_percentiles_symmetry = F
)
}
\arguments{
\item{list_of_lists_of_different_pop}{a list that is mode of lists}

\item{by}{number: increment of the sequence.}

\item{enforce_percentiles_symmetry}{in case a symmetry between the probabilities percentiles is desired}
}
\description{
# In order to create a performance table for different population with different 
outcomes the user should use a list of lists like in the example below:
}
\examples{


create_performance_table_multiple_pop(list(
"train" = list(
 probs = example_dat \%>\%
 dplyr::filter(type_of_set == "train") \%>\%
 dplyr::pull(estimated_probabilities),
 real = example_dat \%>\% dplyr::filter(type_of_set == "train") \%>\%
 dplyr::pull(outcome)),
"test" = list(
 probs = example_dat \%>\% dplyr::filter(type_of_set == "test") \%>\%
 dplyr::pull(estimated_probabilities),
 real = example_dat \%>\% dplyr::filter(type_of_set == "test") \%>\%
 dplyr::pull(outcome))
))


# Each element in the list represent a single population that one vector of predictions
# and another vector of outcomes. 


}
