% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_helper_functions.R
\name{create_plotly_base}
\alias{create_plotly_base}
\title{Create basic plotly for Performance Metrics}
\usage{
create_plotly_base(
  performance_table,
  x_perf_metric,
  y_perf_metric,
  performance_table_type = "one model",
  col_values = c("#5E7F9A", "#931B53", "#F7DC2E", "#C6C174", "#75DBCD")
)
}
\arguments{
\item{performance_table}{an rtichoke performance table}

\item{x_perf_metric}{a performance metrice for the x axis}

\item{y_perf_metric}{a performance metrice for the y axis}

\item{performance_table_type}{the type of the performance table}

\item{col_values}{palette}
}
\description{
Makes a basic plotly for the metrices
}
