% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_plotly_curve}
\alias{create_plotly_curve}
\title{Creating rtichoke plotly curve}
\usage{
create_plotly_curve(rtichoke_curve_list)
}
\arguments{
\item{rtichoke_curve_list}{rtichoke curve list}
}
\description{
Creating rtichoke plotly curve
}
\examples{
\dontrun{

one_pop_one_model |> 
  create_rtichoke_curve_list("roc") |> 
  create_plotly_curve()
  
one_pop_one_model_by_ppcr |> 
  create_rtichoke_curve_list("roc") |> 
  create_plotly_curve()
  
multiple_models |> 
  create_rtichoke_curve_list("roc") |> 
  create_plotly_curve()
  
multiple_models_by_ppcr |> 
  create_rtichoke_curve_list("roc") |> 
  create_plotly_curve()
  
multiple_populations |> 
  create_rtichoke_curve_list("roc") |> 
  create_plotly_curve()
  
multiple_populations_by_ppcr |> 
  create_rtichoke_curve_list("roc") |> 
  create_plotly_curve()
  
}
}
\keyword{internal}
