% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_for_performance_metrics.R
\name{create_plotly_for_performance_metrics}
\alias{create_plotly_for_performance_metrics}
\title{Create plotly for Performance Metrics}
\usage{
create_plotly_for_performance_metrics(
  performance_table,
  x_perf_metric,
  y_perf_metric,
  col_values = c("#5E7F9A", "#931B53", "#F7DC2E", "#C6C174", "#75DBCD"),
  main_slider = "threshold"
)
}
\arguments{
\item{performance_table}{an rtichoke performance table}

\item{x_perf_metric}{a performance metrice for the x axis}

\item{y_perf_metric}{a performance metrice for the y axis}

\item{col_values}{color palette}
}
\description{
Makes a ggplot for the metrices
}
