% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_reference_lines_data_frame}
\alias{create_reference_lines_data_frame}
\title{Creating Reference Lines Data Frame}
\usage{
create_reference_lines_data_frame(
  curve,
  prevalence = NA,
  color = NA,
  plotly = FALSE,
  multiple_pop = FALSE
)
}
\arguments{
\item{curve}{the specified curve for the reference lines}

\item{prevalence}{the prevalence of the outcome}

\item{color}{the required color}

\item{plotly}{should the reference lines data frame be
competible with plotly}

\item{multiple_pop}{should the reference lines data frame should be
adjusted to multiple populations}
}
\description{
Creating Reference Lines Data Frame
}
\keyword{internal}
