% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_table.R
\name{create_subtitle_for_one_population}
\alias{create_subtitle_for_one_population}
\title{Creating Subtitle for One Population}
\usage{
create_subtitle_for_one_population(pop_name, pop_color, pop_prevalence)
}
\arguments{
\item{pop_name}{the name of the population}

\item{pop_color}{the color of the population}

\item{pop_prevalence}{the prevalence of the population}
}
\value{

}
\description{
Creating Subtitle for One Population
}
\keyword{internal}
