% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_text.R
\name{create_text_for_hover}
\alias{create_text_for_hover}
\title{Creating text to hover}
\usage{
create_text_for_hover(
  performance_data_type,
  curve,
  stratified_by = "probability_threshold"
)
}
\arguments{
\item{performance_data_type}{performance data type}

\item{curve}{curve type}

\item{stratified_by}{A character string specifying the stratification method.}
}
\description{
Creating text to hover
}
\keyword{internal}
