% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_table.R
\name{creating_title_for_gt}
\alias{creating_title_for_gt}
\title{Creating Title for gt performance table}
\usage{
creating_title_for_gt(main_slider)
}
\description{
Creating Title for gt performance table
}
\keyword{internal}
