% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{define_limits_for_calibration_plot}
\alias{define_limits_for_calibration_plot}
\title{Define limits for Calibration Curve}
\usage{
define_limits_for_calibration_plot(calibration_bins_dat)
}
\arguments{
\item{calibration_bins_dat}{A data frame containing bin-level calibration data.}
}
\description{
Define limits for Calibration Curve
}
\examples{
\dontrun{
make_calibration_bins_dat(
  probs = example_dat$estimated_probabilities,
  reals = example_dat$outcome
) |>
  define_limits_for_calibration_plot()
}
}
\keyword{internal}
