% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_dat}
\alias{example_dat}
\title{Example Data Frame for rtichoke package}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
  \item{estimated_probabilities}{a vector of estimated probabilities from some model}
  \item{random_guess}{a vector of random guesses}
  \item{outcome}{a vector of binary outcome}
  ...
}
}
\usage{
example_dat
}
\description{
An example data frame that contains estimated probabilities and binary outcomes,
used as an input for rtichoke functions.
}
\keyword{datasets}
