% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_prevalence_from_performance_data}
\alias{get_prevalence_from_performance_data}
\title{prevalence}
\usage{
get_prevalence_from_performance_data(performance_data)
}
\arguments{
\item{performance_data}{an rtichoke Performance Data}
}
\description{
Get the prevalence out of Performance Data
}
\keyword{internal}
