% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_prevalence_from_performance_table}
\alias{get_prevalence_from_performance_table}
\title{prevalence}
\usage{
get_prevalence_from_performance_table(performance_table)
}
\arguments{
\item{performance_table}{an rtichoke performance table}
}
\description{
Get the prevalence out of performance table
}
