% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_real_positives_from_performance_data}
\alias{get_real_positives_from_performance_data}
\title{Real Positives}
\usage{
get_real_positives_from_performance_data(
  performance_data,
  performance_data_type = "not important"
)
}
\arguments{
\item{performance_data}{an rtichoke Performance Data}

\item{performance_data_type}{the type of the Performance Data}
}
\description{
Get the real positives out of Performance Data
}
\keyword{internal}
