% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_for_performance_metrics.R
\name{interactive_marker_list}
\alias{interactive_marker_list}
\title{Title}
\usage{
interactive_marker_list(multiple_models = F, theme_color = NA)
}
\arguments{
\item{multiple_models}{}

\item{theme_color}{}
}
\description{
Title
}
