% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_text.R
\name{make_performance_metrics_bold}
\alias{make_performance_metrics_bold}
\title{Making performance metrics bold}
\usage{
make_performance_metrics_bold(text_for_hover, curve)
}
\arguments{
\item{text_for_hover}{text}

\item{curve}{curve type}
}
\value{

}
\description{
Making performance metrics bold
}
\keyword{internal}
