% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_text.R
\name{make_two_performance_metrics_bold}
\alias{make_two_performance_metrics_bold}
\title{Title}
\usage{
make_two_performance_metrics_bold(
  text_for_hover,
  performance_metric_x,
  performance_metric_y
)
}
\arguments{
\item{text_for_hover}{text}

\item{performance_metric_x}{x}

\item{performance_metric_y}{y}
}
\value{

}
\description{
Title
}
