% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multiple_models}
\alias{multiple_models}
\title{Example Performance Data}
\format{
A data frame with 202 rows and 15 variables:
}
\usage{
multiple_models
}
\description{
An example Performance Data that contains multiple models
}
\keyword{datasets}
