% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{one_pop_one_model_as_a_vector}
\alias{one_pop_one_model_as_a_vector}
\title{Example Performance Data}
\format{
A data frame with 101 rows and 14 variables:
}
\usage{
one_pop_one_model_as_a_vector
}
\description{
An example Performance Data that contains one model
}
\keyword{datasets}
