% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{one_pop_one_model_as_a_vector_enforced_percentiles_symmetry}
\alias{one_pop_one_model_as_a_vector_enforced_percentiles_symmetry}
\title{Example performance table}
\format{
A data frame with 101 rows and 14 variables:
}
\usage{
one_pop_one_model_as_a_vector_enforced_percentiles_symmetry
}
\description{
An example performance table that contains one model
}
\keyword{datasets}
