% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{one_pop_one_model_by_ppcr}
\alias{one_pop_one_model_by_ppcr}
\title{Example Performance Data}
\format{
A data frame with 101 rows and 14 variables:
}
\usage{
one_pop_one_model_by_ppcr
}
\description{
An example Performance Data that contains one model stratified by ppcr
}
\keyword{datasets}
