% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{one_pop_three_models}
\alias{one_pop_three_models}
\title{Example Performance Data}
\format{
A data frame with 303 rows and 15 variables:
}
\usage{
one_pop_three_models
}
\description{
An example Performance Data that contains three models
}
\keyword{datasets}
