% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{performance_table_for_train_and_test_sets}
\alias{performance_table_for_train_and_test_sets}
\title{Example performance table}
\format{
A data frame with 202 rows and 15 variables:
}
\usage{
performance_table_for_train_and_test_sets
}
\description{
An example performance table that contains train and test sets
}
\keyword{datasets}
