% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{performance_table_for_two_models}
\alias{performance_table_for_two_models}
\title{Example performance table}
\format{
A data frame with 202 rows and 15 variables:
}
\usage{
performance_table_for_two_models
}
\description{
An example performance table that contains performance for two different models
}
\keyword{datasets}
