% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{performance_table_one_model}
\alias{performance_table_one_model}
\title{Example performance table}
\format{
A data frame with 101 rows and 14 variables:
}
\usage{
performance_table_one_model
}
\description{
An example performance table that contains one model for one population
}
\keyword{datasets}
