% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision.R
\name{plot_conventional_decision}
\alias{plot_conventional_decision}
\title{Decision Curve from Performance Data (Conventional)}
\usage{
plot_conventional_decision(
  performance_data,
  chosen_threshold = NA,
  interactive = TRUE,
  col_values = c("#5BC0BE", "#FC8D62", "#8DA0CB", "#E78AC3", "#A4243B"),
  size = NULL
)
}
\arguments{
\item{performance_data}{an rtichoke Performance Data}

\item{chosen_threshold}{a chosen threshold to display (for non-interactive)}

\item{interactive}{whether the plot should be interactive
plots}

\item{col_values}{color palette}

\item{size}{the size of the curve}
}
\description{
Plot Conventional Decision Curve
}
\examples{

\dontrun{

one_pop_one_model \%>\%
  rtichoke:::plot_conventional_decision()

multiple_models \%>\%
  rtichoke:::plot_conventional_decision()

multiple_populations \%>\%
  rtichoke:::plot_conventional_decision()
  
}
}
\keyword{internal}
