% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision.R
\name{plot_interventions_avoided}
\alias{plot_interventions_avoided}
\title{Decision Curve from Performance Data (Interventions Avoided)}
\usage{
plot_interventions_avoided(
  performance_data,
  chosen_threshold = NA,
  interactive = TRUE,
  col_values = c("#5BC0BE", "#FC8D62", "#8DA0CB", "#E78AC3", "#A4243B"),
  size = NULL,
  min_p_threshold = 0,
  max_p_threshold = 1
)
}
\arguments{
\item{performance_data}{an rtichoke Performance Data}

\item{chosen_threshold}{a chosen threshold to display (for non-interactive)}

\item{interactive}{whether the plot should be interactive
plots}

\item{col_values}{color palette}

\item{size}{the size of the curve}

\item{min_p_threshold}{The minimum Probability Threshold value to be
displayed}

\item{max_p_threshold}{The maximum Probability Threshold value to be
displayed}
}
\description{
Plot interventions Avoided
}
\examples{

\dontrun{

one_pop_one_model \%>\%
  rtichoke:::plot_interventions_avoided()

multiple_models \%>\%
  rtichoke:::plot_interventions_avoided()

multiple_populations \%>\%
  rtichoke:::plot_interventions_avoided()
  
}
}
\keyword{internal}
