% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_table.R
\name{prepare_performance_data_for_gt}
\alias{prepare_performance_data_for_gt}
\title{Preparing Performance Data for gt}
\usage{
prepare_performance_data_for_gt(performance_data, main_slider)
}
\arguments{
\item{performance_data}{an rtichoke Performance Data}
}
\description{
Preparing Performance Data for gt
}
\keyword{internal}
