% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_for_performance_metrics.R
\name{remove_grid_lines_from_plotly}
\alias{remove_grid_lines_from_plotly}
\title{Title}
\usage{
remove_grid_lines_from_plotly(plot_tichoke_ly_ob)
}
\arguments{
\item{plot_tichoke_ly_ob}{}
}
\description{
Title
}
