% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrix_list.R
\name{render_reactable_confusion_matrix}
\alias{render_reactable_confusion_matrix}
\title{Render a reactable confusion matrix}
\usage{
render_reactable_confusion_matrix(confusion_matrix)
}
\arguments{
\item{confusion_matrix}{a confusion matrix}
}
\description{
Render a reactable confusion matrix
}
\keyword{internal}
