% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_table_gt_functions.R
\name{replace_nan_with_na}
\alias{replace_nan_with_na}
\title{Replancing NaN with NA}
\usage{
replace_nan_with_na(performance_dat)
}
\arguments{
\item{performance_dat}{the original performance data}
}
\value{

}
\description{
Replancing NaN with NA
}
\keyword{internal}
