% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtichoke-package.R
\docType{package}
\name{rtichoke-package}
\alias{rtichoke}
\alias{rtichoke-package}
\title{rtichoke: rtichoke}
\description{
Interactive visualization for model performance.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://uriahf.github.io/rtichoke/}
  \item \url{https://github.com/uriahf/rtichoke}
  \item Report bugs at \url{https://github.com/uriahf/rtichoke/issues}
}

}
\author{
\strong{Maintainer}: Uriah Finkel \email{ufinkel@gmail.com}

Authors:
\itemize{
  \item Uriah Finkel \email{ufinkel@gmail.com}
}

}
\keyword{internal}
