% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_helper_functions.R
\name{set_axis_titles}
\alias{set_axis_titles}
\title{Set Titles for x and y axis in plotly objects}
\usage{
set_axis_titles(plotly_object, curve, max_y_range = NA, min_y_range = NA)
}
\arguments{
\item{plotly_object}{a plotly object}

\item{curve}{the required curve}

\item{max_y_range}{the maximum value for y range}

\item{min_y_range}{the minimum value for y range}
}
\description{
Set Titles for x and y axis in plotly objects
}
\keyword{internal}
