% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision.R
\name{set_decision_curve_limits}
\alias{set_decision_curve_limits}
\title{Set the limits for Decision Curve}
\usage{
set_decision_curve_limits(decision_curve)
}
\arguments{
\item{decision_curve}{a ggplot object of Decision Curve}
}
\description{
Set the limits for Decision Curve
}
\keyword{internal}
