% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gains.R
\name{set_gains_curve_limits}
\alias{set_gains_curve_limits}
\title{Set the limits for Gains Curve}
\usage{
set_gains_curve_limits(gains_curve)
}
\arguments{
\item{gains_curve}{a ggplot object of a Gains Curve}
}
\description{
Set the limits for Gains Curve
}
