% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lift.R
\name{set_lift_curve_limits}
\alias{set_lift_curve_limits}
\title{Set the limits for lift curve}
\usage{
set_lift_curve_limits(lift_curve)
}
\arguments{
\item{lift_curve}{a ggplot object of lift curve}
}
\description{
Set the limits for lift curve
}
