% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_for_performance_metrics.R
\name{set_lift_limits}
\alias{set_lift_limits}
\title{Set the limits for Lift Curve}
\usage{
set_lift_limits(lift_curve)
}
\arguments{
\item{lift_curve}{a ggplot object of a Lift Curve}
}
\description{
Set the limits for Lift Curve
}
