% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_for_performance_metrics.R
\name{set_precision_recall_curve_limits}
\alias{set_precision_recall_curve_limits}
\title{Set the limits for percision recall curve}
\usage{
set_precision_recall_curve_limits(precision_recall_curve)
}
\arguments{
\item{precision_recall_curve}{a ggplot object of precision recall curve}
}
\description{
Set the limits for percision recall curve
}
