% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_helper_functions.R
\name{set_styling_for_rtichoke}
\alias{set_styling_for_rtichoke}
\title{Set styling for rtichoke plotly}
\usage{
set_styling_for_rtichoke(
  plotly_object,
  curve,
  min_y_range = NA,
  max_y_range = NA,
  min_x_range = NA,
  max_x_range = NA
)
}
\arguments{
\item{plotly_object}{a plotly object}

\item{curve}{the required curve}

\item{min_y_range}{the minimum value of y range (for decision curve)}

\item{max_y_range}{the maximum value of y range (for lift and decision curve)}

\item{min_x_range}{the minimum value of x range (for decision curve)}

\item{max_x_range}{the maximum value of x range (for decision curve)}
}
\description{
Set styling for rtichoke plotly
}
\keyword{internal}
