% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{train_and_test_sets}
\alias{train_and_test_sets}
\title{Example performance table}
\format{
A data frame with 202 rows and 15 variables:
}
\usage{
train_and_test_sets
}
\description{
An example performance table that contains test and train sets
}
\keyword{datasets}
