% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{train_and_test_sets_enforced_percentiles_symmetry}
\alias{train_and_test_sets_enforced_percentiles_symmetry}
\title{Example Performance Data}
\format{
A data frame with 202 rows and 15 variables:
}
\usage{
train_and_test_sets_enforced_percentiles_symmetry
}
\description{
An example Performance Data that contains test and train sets
}
\keyword{datasets}
