% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_text.R
\name{add_hover_text_to_performance_data}
\alias{add_hover_text_to_performance_data}
\title{Adding hover text to performance data}
\usage{
add_hover_text_to_performance_data(
  performance_data,
  performance_data_type,
  curve,
  stratified_by = "probability_threshold"
)
}
\arguments{
\item{performance_data}{perf dat}

\item{performance_data_type}{perf dat type}

\item{curve}{curve}

\item{stratified_by}{Performance Metrics can be stratified by Probability
Threshold or alternatively by Predicted Positives Condition Rate}
}
\description{
Adding hover text to performance data
}
\keyword{internal}
