% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_text.R
\name{add_hover_text_to_performance_data}
\alias{add_hover_text_to_performance_data}
\title{Title}
\usage{
add_hover_text_to_performance_data(
  performance_data,
  performance_data_type,
  curve
)
}
\arguments{
\item{performance_data}{perf dat}

\item{performance_data_type}{perf dat type}

\item{curve}{curve}
}
\value{

}
\description{
Title
}
