% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_helper_functions.R
\name{add_interactive_marker_from_performance_data}
\alias{add_interactive_marker_from_performance_data}
\title{Add interactive marker based on performance data}
\usage{
add_interactive_marker_from_performance_data(
  plotly_object,
  performance_data,
  performance_data_type,
  x_perf_metric,
  y_perf_metric,
  main_slider = "threshold"
)
}
\arguments{
\item{plotly_object}{}

\item{performance_data}{}

\item{performance_data_type}{}

\item{x_perf_metric}{}

\item{y_perf_metric}{}

\item{main_slider}{}
}
\value{

}
\description{
Add interactive marker based on performance data
}
