% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_helper_functions.R
\name{add_interactive_marker_from_performance_data}
\alias{add_interactive_marker_from_performance_data}
\title{Add interactive marker based on performance data}
\usage{
add_interactive_marker_from_performance_data(
  plotly_object,
  performance_data,
  performance_data_type,
  x_perf_metric,
  y_perf_metric,
  main_slider = "threshold"
)
}
\arguments{
\item{plotly_object}{a previous plotly object}

\item{performance_data}{the performance data for the plot}

\item{performance_data_type}{the type of the performance data}

\item{x_perf_metric}{performance metric for the x axis}

\item{y_perf_metric}{performance metric for the y axis}

\item{main_slider}{the main slider for interactivity}
}
\value{

}
\description{
Add interactive marker based on performance data
}
\keyword{internal}
