% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly helper functions.R
\name{add_interactive_marker_to_plotly}
\alias{add_interactive_marker_to_plotly}
\title{Add Interactive Marker to Plotly}
\usage{
add_interactive_marker_to_plotly(
  plotly_object,
  performance_table_type,
  main_slider = "threshold"
)
}
\arguments{
\item{main_slider}{}

\item{plot_tichoke_ly_ob}{}
}
\description{
Add Interactive Marker to Plotly
}
