% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_helper_functions.R
\name{add_interactive_marker_to_plotly}
\alias{add_interactive_marker_to_plotly}
\title{Add Interactive Marker to Plotly}
\usage{
add_interactive_marker_to_plotly(plotly_object, main_slider = "threshold")
}
\arguments{
\item{plotly_object}{a plotly plot for performance metrics}

\item{main_slider}{what is the main slider - threshold, percent positives or positives}
}
\description{
Add Interactive Marker to Plotly
}
