% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_helper_functions.R
\name{add_lines_and_markers_from_performance_data}
\alias{add_lines_and_markers_from_performance_data}
\title{Add lines and markers based on performance data}
\usage{
add_lines_and_markers_from_performance_data(
  plotly_reference_lines,
  performance_data,
  performance_data_type,
  x_perf_metric,
  y_perf_metric,
  col_values = c("#5E7F9A", "#931B53", "#F7DC2E", "#C6C174", "#75DBCD"),
  main_slider = "threshold",
  reference_lines = NA
)
}
\arguments{
\item{plotly_reference_lines}{}

\item{performance_data}{}

\item{performance_data_type}{}

\item{x_perf_metric}{}

\item{y_perf_metric}{}

\item{col_values}{}

\item{main_slider}{}

\item{reference_lines}{}
}
\value{

}
\description{
Add lines and markers based on performance data
}
