% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_helper_functions.R
\name{add_lines_and_markers_from_performance_data}
\alias{add_lines_and_markers_from_performance_data}
\title{Add lines and markers based on performance data}
\usage{
add_lines_and_markers_from_performance_data(
  plotly_object,
  performance_data,
  performance_data_type,
  x_perf_metric,
  y_perf_metric,
  col_values = c("#5BC0BE", "#FC8D62", "#8DA0CB", "#E78AC3", "#A4243B"),
  main_slider = "threshold"
)
}
\arguments{
\item{plotly_object}{a previous plotly object}

\item{performance_data}{the performance data for the plot}

\item{performance_data_type}{the type of the performance data}

\item{x_perf_metric}{performance metric for the x axis}

\item{y_perf_metric}{performance metric for the y axis}

\item{col_values}{color palette}

\item{main_slider}{the main slider for interactivity}
}
\value{

}
\description{
Add lines and markers based on performance data
}
\keyword{internal}
