% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_text.R
\name{add_models_for_text_for_hover}
\alias{add_models_for_text_for_hover}
\title{Adding models for text to hover}
\usage{
add_models_for_text_for_hover(text_for_hover)
}
\arguments{
\item{text_for_hover}{text}
}
\value{

}
\description{
Adding models for text to hover
}
\keyword{internal}
