% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_for_performance_metrics.R
\name{add_prevalence_layers_to_gains_curve}
\alias{add_prevalence_layers_to_gains_curve}
\title{Add Prevalence Layers go Gains Curve}
\usage{
add_prevalence_layers_to_gains_curve(prevalence, col_value)
}
\arguments{
\item{prevalence}{the prevalence in each population}

\item{col_value}{color values palette}
}
\description{
Add Prevalence Layers go Gains Curve
}
