% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_helper_functions.R
\name{add_reference_lines_to_plotly}
\alias{add_reference_lines_to_plotly}
\title{Add Reference Lines to Plotly Object}
\usage{
add_reference_lines_to_plotly(
  plotly_object,
  reference_lines,
  performance_table_type = "one model"
)
}
\arguments{
\item{plotly_object}{a plotly plot for performance metrics}

\item{reference_lines}{dataframe of reference lines}

\item{performance_table_type}{the type of the performance table}
}
\description{
Add Reference Lines to Plotly Object
}
