% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{arrange_estimated_probabilities_to_long_format}
\alias{arrange_estimated_probabilities_to_long_format}
\title{Arrange estimated probabilities to long format}
\usage{
arrange_estimated_probabilities_to_long_format(probs)
}
\arguments{
\item{probs}{a vector of estimated probabilities or a list of vectors of that
kind (one for each model)}
}
\value{

}
\description{
Arrange estimated probabilities to long format
}
\keyword{)}
\keyword{=}
\keyword{Model"}
\keyword{\dontrun{}
\keyword{arrange_estimated_probabilities_to_long_format(}
\keyword{example_dat$estimated_probabilities)}
\keyword{internal}
\keyword{list("First}
\keyword{probs}
\keyword{}}
