% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{arrange_estimated_probabilities_to_long_format}
\alias{arrange_estimated_probabilities_to_long_format}
\title{Arrange estimated probabilities to long format}
\usage{
arrange_estimated_probabilities_to_long_format(probs, real)
}
\arguments{
\item{probs}{a vector of estimated probabilities or a list of vectors of that
kind (one for each model)}

\item{real}{a vector of binary outcomes or a list of vectors of that
kind (one for each population)}
}
\value{

}
\description{
Arrange estimated probabilities to long format
}
\examples{
arrange_estimated_probabilities_to_long_format(
probs = list("First Model" = example_dat$estimated_probabilities),
real = example_dat$outcome
)
}
\keyword{internal}
