% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_table.R
\name{bar_style_nb_reactable}
\alias{bar_style_nb_reactable}
\title{Add background color}
\usage{
bar_style_nb_reactable(width = 1, pos_fill = "lightgreen", neg_fill = "pink")
}
\arguments{
\item{The}{width of the background color}
}
\description{
Add background color
}
\keyword{internal}
