% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_table.R
\name{bar_style_perf}
\alias{bar_style_perf}
\title{Add background color}
\usage{
bar_style_perf(width = 1, color = "lightgreen")
}
\arguments{
\item{The}{width of the background color}
}
\description{
Add background color
}
\keyword{internal}
