% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs.R
\name{check_real_input}
\alias{check_real_input}
\title{Check real input}
\usage{
check_real_input(real)
}
\description{
Check real input
}
\examples{
\dontrun{
check_real_input(example_dat$outcome)

list(
  "train" = example_dat \%>\%
    dplyr::filter(type_of_set == "train") \%>\%
    dplyr::pull(outcome),
  "test" = example_dat \%>\%
    dplyr::filter(type_of_set == "test") \%>\%
    dplyr::pull(outcome)
) \%>\%
  check_real_input()

check_real_input(c(example_dat$outcome, -0.1))
check_real_input(c(example_dat$outcome, 1.1))

list(
  "train" = example_dat \%>\%
    dplyr::filter(type_of_set == "train") \%>\%
    dplyr::pull(outcome),
  "test" = c(example_dat \%>\% dplyr::filter(type_of_set == "test") \%>\%
    dplyr::pull(outcome), -0.2)
) \%>\%
  check_real_input()
}
}
\keyword{internal}
