% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{create_calibration_curve}
\alias{create_calibration_curve}
\title{Create a Calibration Curve}
\usage{
create_calibration_curve(
  probs,
  reals,
  interactive = TRUE,
  col_values = c("#5BC0BE", "#FC8D62", "#8DA0CB", "#E78AC3", "#A4243B"),
  type = "discrete",
  size = NULL
)
}
\arguments{
\item{probs}{a list of vectors of estimated probabilities
(one for each model or one for each population)}

\item{reals}{a list of vectors of binary outcomes (one for each population)}

\item{interactive}{whether the plot should be interactive
plots}

\item{col_values}{color palette}

\item{type}{discrete or smooth}

\item{size}{the size of the curve}
}
\description{
Create a Calibration Curve
}
\examples{
\dontrun{

create_calibration_curve(
  probs = list(example_dat$estimated_probabilities),
  reals = list(example_dat$outcome), type = "discrete"
)


create_calibration_curve(
  probs = list(example_dat$estimated_probabilities),
  reals = list(example_dat$outcome), type = "smooth"
)

# Several Models

create_calibration_curve(
  probs = list(
    "First Model" = example_dat$estimated_probabilities,
    "Second Model" = example_dat$random_guess
  ),
  reals = list(example_dat$outcome), 
  type = "discrete"
)


create_calibration_curve(
  probs = list(
    "First Model" = example_dat$estimated_probabilities,
    "Second Model" = example_dat$random_guess
  ),
  reals = list(example_dat$outcome),
  type = "smooth"
)


# Several Populations


create_calibration_curve(
  probs = list(
    "train" = example_dat \%>\%
      dplyr::filter(type_of_set == "train") \%>\%
      dplyr::pull(estimated_probabilities),
    "test" = example_dat \%>\% dplyr::filter(type_of_set == "test") \%>\%
      dplyr::pull(estimated_probabilities)
  ),
  reals = list(
    "train" = example_dat \%>\% dplyr::filter(type_of_set == "train") \%>\%
      dplyr::pull(outcome),
    "test" = example_dat \%>\% dplyr::filter(type_of_set == "test") \%>\%
      dplyr::pull(outcome)
  ), 
  type = "discrete"
)


create_calibration_curve(
  probs = list(
    "train" = example_dat \%>\%
      dplyr::filter(type_of_set == "train") \%>\%
      dplyr::pull(estimated_probabilities),
    "test" = example_dat \%>\% dplyr::filter(type_of_set == "test") \%>\%
      dplyr::pull(estimated_probabilities)
  ),
  reals = list(
    "train" = example_dat \%>\% dplyr::filter(type_of_set == "train") \%>\%
      dplyr::pull(outcome),
    "test" = example_dat \%>\% dplyr::filter(type_of_set == "test") \%>\%
      dplyr::pull(outcome)
  ), 
  type = "smooth"
)

}


}
