% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{create_calibration_curve}
\alias{create_calibration_curve}
\title{Create a Calibration Curve}
\usage{
create_calibration_curve(
  probs,
  real,
  col_values = c("#5BC0BE", "#FC8D62", "#8DA0CB", "#E78AC3", "#A4243B"),
  type = "discrete"
)
}
\arguments{
\item{probs}{a vector of estimated probabilities or a list of vectors of that
kind (one for each model)}

\item{real}{a vector of binary outcomes or a list of vectors of that
kind (one for each population)}

\item{col_values}{color palette}
}
\value{

}
\description{
Create a Calibration Curve
}
\examples{
create_calibration_curve(
  probs = example_dat$estimated_probabilities,
  real = example_dat$outcome
)

create_calibration_curve(
  probs = list(
    "First Model" = example_dat$estimated_probabilities,
    "Second Model" = example_dat$random_guess
  ),
  real = example_dat$outcome
)

\dontrun{
create_calibration_curve(
  probs = list(
    "train" = example_dat \%>\%
      dplyr::filter(type_of_set == "train") \%>\%
      dplyr::pull(estimated_probabilities),
    "test" = example_dat \%>\% dplyr::filter(type_of_set == "test") \%>\%
      dplyr::pull(estimated_probabilities)
  ),
  real = list(
    "train" = example_dat \%>\% dplyr::filter(type_of_set == "train") \%>\%
      dplyr::pull(outcome),
    "test" = example_dat \%>\% dplyr::filter(type_of_set == "test") \%>\%
      dplyr::pull(outcome)
  )
)
}
}
