% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrix_list.R
\name{create_conf_mat_list}
\alias{create_conf_mat_list}
\title{Create a list of confusion matrices}
\usage{
create_conf_mat_list(
  performance_table,
  stratified_by = "probability_threshold"
)
}
\description{
Create a list of confusion matrices
}
\examples{
\dontrun{

one_pop_one_model \%>\%
  create_conf_mat_list()

one_pop_one_model_by_ppcr \%>\%
  create_conf_mat_list()

multiple_models \%>\%
  create_conf_mat_list()

multiple_models_by_ppcr \%>\%
  create_conf_mat_list()

multiple_populations \%>\%
  create_conf_mat_list()

multiple_populations_by_ppcr \%>\%
  create_conf_mat_list()
}
}
\keyword{internal}
