% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrix_list.R
\name{create_conf_mat_list}
\alias{create_conf_mat_list}
\title{Create a list of confusion matrices}
\usage{
create_conf_mat_list(performance_table, main_slider = "threshold")
}
\description{
Create a list of confusion matrices
}
\examples{
\dontrun{
one_pop_one_model_as_a_vector \%>\%
  create_conf_mat_list()

one_pop_one_model_as_a_vector_enforced_percentiles_symmetry \%>\%
  create_conf_mat_list(main_slider = "ppcr")

one_pop_one_model_as_a_list \%>\%
  create_conf_mat_list()

one_pop_one_model_as_a_list_enforced_percentiles_symmetry \%>\%
  create_conf_mat_list(main_slider = "ppcr")

one_pop_three_models \%>\%
  create_conf_mat_list()

one_pop_three_models_enforced_percentiles_symmetry \%>\%
  create_conf_mat_list(main_slider = "ppcr")

train_and_test_sets \%>\%
  create_conf_mat_list()

train_and_test_sets_enforced_percentiles_symmetry \%>\%
  create_conf_mat_list(main_slider = "ppcr")

one_pop_one_model_as_a_vector \%>\%
  create_conf_mat_list()

one_pop_one_model_as_a_vector_enforced_percentiles_symmetry \%>\%
  create_conf_mat_list(main_slider = "ppcr")
}
}
\keyword{internal}
