% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{create_dat_for_smooth_calibration}
\alias{create_dat_for_smooth_calibration}
\title{Creating dat for smooth calibration}
\usage{
create_dat_for_smooth_calibration(probs, reals, deciles_dat)
}
\arguments{
\item{probs}{a list of vectors of estimated probabilities
(one for each model or one for each population)}

\item{reals}{a list of vectors of binary outcomes (one for each population)}

\item{deciles_dat}{data of deciles}
}
\description{
Arrange estimated probabilities to long format
}
\examples{
\dontrun{
# several models
deciles_dat <- purrr::map_df(list(
  "Model 1" = example_dat$estimated_probabilities
),
~ make_deciles_dat(.x, example_dat$outcome),
.id = "model"
) \%>\%
  mutate(model = forcats::fct_inorder(factor(model)))

create_dat_for_smooth_calibration(
  list("Model 1" = example_dat$estimated_probabilities),
  reals = list(example_dat$outcome),
  deciles_dat
)

# several populations

deciles_dat <- purrr::map2_dfr(list(
  "train" = example_dat \%>\%
    dplyr::filter(type_of_set == "train") \%>\%
    dplyr::pull(estimated_probabilities),
  "test" = example_dat \%>\% dplyr::filter(type_of_set == "test") \%>\%
    dplyr::pull(estimated_probabilities)
),
list(
  "train" = example_dat \%>\% dplyr::filter(type_of_set == "train") \%>\%
    dplyr::pull(outcome),
  "test" = example_dat \%>\% dplyr::filter(type_of_set == "test") \%>\%
    dplyr::pull(outcome)
),
~ make_deciles_dat(.x, .y),
.id = "population"
) \%>\%
  mutate(population = forcats::fct_inorder(factor(population)))

create_dat_for_smooth_calibration(
  probs = list(
    "train" = example_dat \%>\%
      dplyr::filter(type_of_set == "train") \%>\%
      dplyr::pull(estimated_probabilities),
    "test" = example_dat \%>\% dplyr::filter(type_of_set == "test") \%>\%
      dplyr::pull(estimated_probabilities)
  ),
  reals = list(
    "train" = example_dat \%>\% dplyr::filter(type_of_set == "train") \%>\%
      dplyr::pull(outcome),
    "test" = example_dat \%>\% dplyr::filter(type_of_set == "test") \%>\%
      dplyr::pull(outcome)
  ),
  deciles_dat
)
}
}
\keyword{internal}
