% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision.R
\name{create_decision_curve}
\alias{create_decision_curve}
\title{Decision Curve}
\usage{
create_decision_curve(
  probs,
  real,
  by = 0.01,
  stratified_by = "probability_threshold"
)
}
\arguments{
\item{probs}{a vector of estimated probabilities or a list of vectors of that
kind (one for each model)}

\item{real}{a vector of binary outcomes or a list of vectors of that
kind (one for each population)}

\item{by}{number: increment of the sequence.}

\item{stratified_by}{Performance Metrics can be stratified by Probability
Threshold or alternatively by Predicted Positives Condition Rate}
}
\description{
Create a Decision Curve
}
