% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_for_performance_metrics.R
\name{create_ggplot_for_performance_metrics}
\alias{create_ggplot_for_performance_metrics}
\title{Create ggplot for Performance Metrics}
\usage{
create_ggplot_for_performance_metrics(
  performance_data,
  x_perf_metric,
  y_perf_metric,
  col_values = c("#5E7F9A", "#931B53", "#F7DC2E", "#C6C174", "#75DBCD")
)
}
\arguments{
\item{performance_data}{an rtichoke Performance Data}

\item{x_perf_metric}{a performance metrice for the x axis}

\item{y_perf_metric}{a performance metrice for the y axis}

\item{col_values}{color palette}
}
\description{
Makes a ggplot for the metrices
}
\keyword{internal}
