% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_for_performance_metrics.R
\name{create_lift_curve}
\alias{create_lift_curve}
\title{Lift Curve}
\usage{
create_lift_curve(probs, real, by = 0.01, enforce_percentiles_symmetry = F)
}
\arguments{
\item{probs}{a vector of estimated probabilities or a list of vectors of that
kind (one for each model)}

\item{real}{a vector of binary outcomes or a list of vectors of that
kind (one for each population)}

\item{by}{number: increment of the sequence.}

\item{enforce_percentiles_symmetry}{in case a symmetry between the probabilities percentiles is desired}
}
\description{
Create a Lift Curve
}
