% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lift_curve.R
\name{create_lift_curve}
\alias{create_lift_curve}
\title{LIFT Curve}
\usage{
create_lift_curve(
  probs,
  real,
  by = 0.01,
  enforce_percentiles_symmetry = F,
  col_values = c("#21DACD", "#B6C174", "#A7DA2E", "#C2C172", "#FFD700")
)
}
\arguments{
\item{probs}{a vector of estimated probabilities or a list of vectors of that
kind (one for each model)}

\item{real}{a vector of binary outcomes or a list of vectors of that
kind (one for each population)}

\item{by}{number: increment of the sequence.}

\item{enforce_percentiles_symmetry}{in case a symmetry between the probabilities percentiles is desired}

\item{col_values}{color palette}
}
\description{
Create a LIFT Curve
}
