% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_table.R
\name{create_performance_table}
\alias{create_performance_table}
\title{Performance Table}
\usage{
create_performance_table(
  probs,
  reals,
  by = 0.01,
  stratified_by = "probability_threshold",
  output_type = "reactable"
)
}
\arguments{
\item{probs}{a list of vectors of estimated probabilities
(one for each model or one for each population)}

\item{reals}{a list of vectors of binary outcomes (one for each population)}

\item{by}{number: increment of the sequence.}

\item{stratified_by}{Performance Metrics can be stratified by Probability
Threshold or alternatively by Predicted Positives Condition Rate}

\item{output_type}{the type of the output table, {gt} as default}
}
\description{
Create a Performance Table
}
