% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_performance_table.R
\name{create_performance_table}
\alias{create_performance_table}
\title{Create Performance Table
Makes a Performance Table}
\usage{
create_performance_table(
  probs,
  real,
  by = 0.01,
  enforce_riskpercentiles_symmetry = F
)
}
\arguments{
\item{probs}{a vector of estimated probabilities or a list if vectors of that kind (one for each model)}

\item{real}{a vector of binaric outcomes}

\item{by}{number: increment of the sequence.}

\item{enforce_riskpercentiles_symmetry}{in case a symmetry between the risk percentiles is desired}
}
\description{
Create Performance Table
Makes a Performance Table
}
