% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_table.R
\name{create_performance_table}
\alias{create_performance_table}
\title{Performance Table}
\usage{
create_performance_table(
  probs,
  real,
  by = 0.01,
  enforce_percentiles_symmetry = F
)
}
\arguments{
\item{probs}{a vector of estimated probabilities or a list of vectors of that
kind (one for each model)}

\item{real}{a vector of binary outcomes or a list of vectors of that
kind (one for each population)}

\item{by}{number: increment of the sequence.}

\item{enforce_percentiles_symmetry}{in case a symmetry between the probabilities percentiles is desired}
}
\description{
Create a Performance Table
}
