% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_helper_functions.R
\name{create_plotly_base}
\alias{create_plotly_base}
\title{Create basic plotly for Performance Metrics}
\usage{
create_plotly_base(
  performance_data,
  x_perf_metric,
  y_perf_metric,
  performance_data_type = "one model",
  col_values = c("#5BC0BE", "#FC8D62", "#8DA0CB", "#E78AC3", "#A4243B"),
  height = size,
  width = size
)
}
\arguments{
\item{performance_data}{an rtichoke Performance Data}

\item{x_perf_metric}{a performance metrice for the x axis}

\item{y_perf_metric}{a performance metrice for the y axis}

\item{performance_data_type}{the type of the Performance Data}

\item{col_values}{palette}
}
\description{
Makes a basic plotly for the metrices
}
\keyword{internal}
