% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision_recall.R
\name{create_precision_recall_curve}
\alias{create_precision_recall_curve}
\title{Precision Recall Curve}
\usage{
create_precision_recall_curve(
  probs,
  real,
  by = 0.01,
  enforce_percentiles_symmetry = F
)
}
\arguments{
\item{probs}{a vector of estimated probabilities or a list of vectors of that
kind (one for each model)}

\item{real}{a vector of binary outcomes or a list of vectors of that
kind (one for each population)}

\item{by}{number: increment of the sequence.}

\item{enforce_percentiles_symmetry}{in case a symmetry between the probabilities percentiles is desired}
}
\description{
Create a Precision Recall Curve
}
