% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_reference_lines_data_frame}
\alias{create_reference_lines_data_frame}
\title{Title}
\usage{
create_reference_lines_data_frame(
  curve,
  prevalence = NA,
  color = NA,
  plotly = F,
  multiple_pop = F
)
}
\arguments{
\item{curve}{the specified curve for the reference lines}

\item{prevalence}{the prevalence of the outcome}

\item{color}{the required color}
}
\description{
Title
}
