% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_helper_functions.R
\name{create_reference_lines_for_plotly}
\alias{create_reference_lines_for_plotly}
\title{Create reference lines plotly as the first stage of interactive plot}
\usage{
create_reference_lines_for_plotly(
  performance_table_type,
  curve,
  prevalence = NA,
  population_color_vector = NA
)
}
\arguments{
\item{performance_table_type}{the type of the performance table}

\item{curve}{the required curve}

\item{prevalence}{the prevalence of the population}

\item{population_color_vector}{color values}
}
\description{
Create reference lines plotly as the first stage of interactive plot
}
\keyword{internal}
