% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R, R/tidymodels.R
\name{create_roc_curve}
\alias{create_roc_curve}
\alias{create_roc_curve.default}
\alias{create_roc_curve.tune_results}
\title{Create a ROC Curve}
\usage{
create_roc_curve(x, ...)

\method{create_roc_curve}{default}(
  x,
  reals,
  by = 0.01,
  stratified_by = "probability_threshold",
  chosen_threshold = NA,
  interactive = TRUE,
  color_values = c("#1b9e77", "#d95f02", "#7570b3", "#e7298a", "#07004D", "#E6AB02",
    "#FE5F55", "#54494B", "#006E90", "#BC96E6", "#52050A", "#1F271B", "#BE7C4D",
    "#63768D", "#08A045", "#320A28", "#82FF9E", "#2176FF", "#D1603D", "#585123"),
  title_included = FALSE,
  size = NULL,
  ...
)

\method{create_roc_curve}{tune_results}(x, ...)
}
\arguments{
\item{x}{An object containing prediction and outcome data.}

\item{...}{Additional arguments passed to specific methods.}

\item{reals}{A list of numeric vectors containing the binary outcomes (0 or 1).}

\item{by}{The increment for the probability threshold sequence.}

\item{stratified_by}{A character string specifying the stratification method.}

\item{chosen_threshold}{A numeric value for a specific threshold to highlight.}

\item{interactive}{A logical indicating whether to create an interactive \code{plotly} chart.}

\item{color_values}{A character vector of hex codes for the color palette.}

\item{title_included}{A logical indicating whether to include a title.}

\item{size}{A numeric value for the line size.}
}
\value{
A \code{plotly} or \code{ggplot2} object.
}
\description{
This is a generic function to create an interactive or static ROC curve
to visualize model performance for binary classification problems.
}
\examples{
\dontrun{
create_roc_curve(
  x = list(example_dat$estimated_probabilities),
  reals = list(example_dat$outcome)
)

create_roc_curve(
  x = list(example_dat$estimated_probabilities),
  reals = list(example_dat$outcome),
  stratified_by = "ppcr"
)

create_roc_curve(
  x = list(
    "First Model" = example_dat$estimated_probabilities,
    "Second Model" = example_dat$random_guess
  ),
  reals = list(example_dat$outcome)
)


create_roc_curve(
  x = list(
    "First Model" = example_dat$estimated_probabilities,
    "Second Model" = example_dat$random_guess
  ),
  reals = list(example_dat$outcome),
  stratified_by = "ppcr"
)


create_roc_curve(
  x = list(
    "train" = example_dat \%>\%
      dplyr::filter(type_of_set == "train") \%>\%
      dplyr::pull(estimated_probabilities),
    "test" = example_dat \%>\% dplyr::filter(type_of_set == "test") \%>\%
      dplyr::pull(estimated_probabilities)
  ),
  reals = list(
    "train" = example_dat \%>\% dplyr::filter(type_of_set == "train") \%>\%
      dplyr::pull(outcome),
    "test" = example_dat \%>\% dplyr::filter(type_of_set == "test") \%>\%
      dplyr::pull(outcome)
  )
)

create_roc_curve(
  x = list(
    "train" = example_dat \%>\%
      dplyr::filter(type_of_set == "train") \%>\%
      dplyr::pull(estimated_probabilities),
    "test" = example_dat \%>\% dplyr::filter(type_of_set == "test") \%>\%
      dplyr::pull(estimated_probabilities)
  ),
  reals = list(
    "train" = example_dat \%>\% dplyr::filter(type_of_set == "train") \%>\%
      dplyr::pull(outcome),
    "test" = example_dat \%>\% dplyr::filter(type_of_set == "test") \%>\%
      dplyr::pull(outcome)
  ),
  stratified_by = "ppcr"
)
}
}
