% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{create_roc_curve}
\alias{create_roc_curve}
\title{ROC Curve}
\usage{
create_roc_curve(
  probs,
  real,
  by = 0.01,
  enforce_percentiles_symmetry = F,
  col_values = c("#5BC0BE", "#FC8D62", "#8DA0CB", "#E78AC3", "#A4243B")
)
}
\arguments{
\item{probs}{a vector of estimated probabilities or a list of vectors of that
kind (one for each model)}

\item{real}{a vector of binary outcomes or a list of vectors of that
kind (one for each population)}

\item{by}{number: increment of the sequence.}

\item{enforce_percentiles_symmetry}{in case a symmetry between the probabilities percentiles is desired}

\item{col_values}{color palette}
}
\description{
Create a ROC Curve
}
