% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_summary_report.R
\name{create_summary_report}
\alias{create_summary_report}
\title{Create a summary report}
\usage{
create_summary_report(
  probs,
  real,
  interactive = FALSE,
  output_file = "summary_report.html"
)
}
\arguments{
\item{probs}{a vector of estimated probabilities or a list of vectors of that
kind (one for each model)}

\item{real}{a vector of binary outcomes or a list of vectors of that
kind (one for each population)}

\item{interactive}{whether the plot should be interactive
plots}

\item{output_file}{The name of the output file}
}
\value{

}
\description{
Create a summary report
}
\examples{
\dontrun{
create_summary_report(
  probs = example_dat$estimated_probabilities,
  real = example_dat$outcome
)

create_summary_report(
  probs = list(
    "First Model" = example_dat$estimated_probabilities,
    "Second Model" = example_dat$random_guess
  ),
  real = example_dat$outcome
)

create_summary_report(
  probs = list(
    "train" = example_dat \%>\%
      dplyr::filter(type_of_set == "train") \%>\%
      dplyr::pull(estimated_probabilities),
    "test" = example_dat \%>\% dplyr::filter(type_of_set == "test") \%>\%
      dplyr::pull(estimated_probabilities)
  ),
  real = list(
    "train" = example_dat \%>\% dplyr::filter(type_of_set == "train") \%>\%
      dplyr::pull(outcome),
    "test" = example_dat \%>\% dplyr::filter(type_of_set == "test") \%>\%
      dplyr::pull(outcome)
  )
)

create_summary_report(
  probs = example_dat$estimated_probabilities,
  real = example_dat$outcome,
  interactive = TRUE
)

create_summary_report(
  probs = list(
    "First Model" = example_dat$estimated_probabilities,
    "Second Model" = example_dat$random_guess
  ),
  real = example_dat$outcome,
  interactive = TRUE
)

create_summary_report(
  probs = list(
    "train" = example_dat \%>\%
      dplyr::filter(type_of_set == "train") \%>\%
      dplyr::pull(estimated_probabilities),
    "test" = example_dat \%>\% dplyr::filter(type_of_set == "test") \%>\%
      dplyr::pull(estimated_probabilities)
  ),
  real = list(
    "train" = example_dat \%>\% dplyr::filter(type_of_set == "train") \%>\%
      dplyr::pull(outcome),
    "test" = example_dat \%>\% dplyr::filter(type_of_set == "test") \%>\%
      dplyr::pull(outcome)
  ),
  interactive = TRUE
)
}
}
