% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_summary_report.R
\name{create_summary_report}
\alias{create_summary_report}
\title{Create a summary report}
\usage{
create_summary_report(
  probs,
  reals,
  interactive = TRUE,
  output_file = "summary_report.html",
  output_dir = getwd()
)
}
\arguments{
\item{probs}{a list of vectors of estimated probabilities
(one for each model or one for each population)}

\item{reals}{a list of vectors of binary outcomes (one for each population)}

\item{interactive}{whether the plot should be interactive
plots}

\item{output_file}{The name of the output file}

\item{output_dir}{The output directory for the rendered \code{output_file}.
This allows for a choice of an alternate directory to which the output file
should be written (the default output directory of that of the input file).
If a path is provided with a filename in \code{output_file} the directory
specified here will take precedence. Please note that any directory path
provided will create any necessary directories if they do not exist.}
}
\description{
Create a summary report
}
\examples{
\dontrun{
create_summary_report(
  probs = list(example_dat$estimated_probabilities),
  reals = list(example_dat$outcome)
)

create_summary_report(
  probs = list(
    "First Model" = example_dat$estimated_probabilities,
    "Second Model" = example_dat$random_guess
  ),
  reals = list(example_dat$outcome)
)

create_summary_report(
  probs = list(
    "train" = example_dat \%>\%
      dplyr::filter(type_of_set == "train") \%>\%
      dplyr::pull(estimated_probabilities),
    "test" = example_dat \%>\% dplyr::filter(type_of_set == "test") \%>\%
      dplyr::pull(estimated_probabilities)
  ),
  reals = list(
    "train" = example_dat \%>\% dplyr::filter(type_of_set == "train") \%>\%
      dplyr::pull(outcome),
    "test" = example_dat \%>\% dplyr::filter(type_of_set == "test") \%>\%
      dplyr::pull(outcome)
  )
)

}
}
