% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_table.R
\name{create_table_for_auc}
\alias{create_table_for_auc}
\title{Create Table for AUC}
\usage{
create_table_for_auc(probs, real)
}
\arguments{
\item{probs}{a vector of estimated probabilities or a list of vectors of that
kind (one for each model)}

\item{real}{a vector of binary outcomes or a list of vectors of that
kind (one for each population)}
}
\description{
Create Table for AUC
}
\keyword{internal}
