% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_table.R
\name{create_table_for_auc}
\alias{create_table_for_auc}
\title{Create Table for AUC}
\usage{
create_table_for_auc(probs, real)
}
\arguments{
\item{probs}{a list of vectors of estimated probabilities
(one for each model or one for each population)}
}
\description{
Create Table for AUC
}
\keyword{internal}
