% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_table.R
\name{create_table_for_prevalence}
\alias{create_table_for_prevalence}
\title{Create Data for Prevalence}
\usage{
create_table_for_prevalence(
  performance_data,
  color_values = c("#1b9e77", "#d95f02", "#7570b3", "#e7298a", "#07004D", "#E6AB02",
    "#FE5F55", "#54494B", "#006E90", "#BC96E6", "#52050A", "#1F271B", "#BE7C4D",
    "#63768D", "#08A045", "#320A28", "#82FF9E", "#2176FF", "#D1603D", "#585123")
)
}
\arguments{
\item{performance_data}{an rtichoke Performance Data}

\item{color_values}{color palette}
}
\description{
Create Data for Prevalence
}
\keyword{internal}
