% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_table.R
\name{create_table_for_prevalence}
\alias{create_table_for_prevalence}
\title{Create Data for Prevalence}
\usage{
create_table_for_prevalence(performance_data)
}
\arguments{
\item{performance_data}{an rtichoke Performance Data}
}
\description{
Create Data for Prevalence
}
\keyword{internal}
