% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_table.R
\name{creating_subtitle_for_gt}
\alias{creating_subtitle_for_gt}
\title{Creating Subtitle for gt performance table}
\usage{
creating_subtitle_for_gt(perf_dat_type, color_values = NA, prevalence = NA)
}
\arguments{
\item{color_values}{color palette}
}
\description{
Creating Subtitle for gt performance table
}
\keyword{internal}
