% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{define_limits_for_calibration_plot}
\alias{define_limits_for_calibration_plot}
\title{Define limits for Calibration Curve}
\usage{
define_limits_for_calibration_plot(deciles_dat)
}
\arguments{
\item{deciles_dat}{}
}
\value{

}
\description{
Define limits for Calibration Curve
}
\examples{
\dontrun{
make_deciles_dat(
probs = example_dat$estimated_probabilities,
real = example_dat$outcome
) \%>\% 
define_limits_for_calibration_plot()
}
}
\keyword{internal}
