% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_dat}
\alias{example_dat}
\title{Example Data Frame for rtichoke package}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
\item{estimated_probabilities}{estimated probabilities from some model}
\item{bad_model}{estimated probabilities from some bad model}
\item{random_guess}{random guesses generated from unifrom distribution}
\item{outcome}{binary outcome}
\item{type_of_set}{the type of the set that the observation belongs to}
...
}
}
\usage{
example_dat
}
\description{
An example data frame that contains estimated probabilities and binary
outcomes, used as an input for rtichoke functions.
}
\keyword{datasets}
