% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_prevalence_from_performance_table}
\alias{get_prevalence_from_performance_table}
\title{prevalence}
\usage{
get_prevalence_from_performance_table(
  performance_table,
  performance_table_type
)
}
\arguments{
\item{performance_table}{an rtichoke performance table}

\item{performance_table_type}{the type of the performance table}
}
\description{
Get the prevalence out of performance table
}
