% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{make_deciles_dat}
\alias{make_deciles_dat}
\title{Make deciles dataframe}
\usage{
make_deciles_dat(probs, real)
}
\arguments{
\item{probs}{a vector of estimated probabilities or a list of vectors of that
kind (one for each model)}

\item{real}{a vector of binary outcomes or a list of vectors of that
kind (one for each population)}
}
\description{
Make deciles dataframe
}
\examples{
\dontrun{
make_deciles_dat(
  probs = example_dat$estimated_probabilities,
  real = example_dat$outcome
)
}
}
\keyword{internal}
