% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_text.R
\name{make_performance_metric_bold}
\alias{make_performance_metric_bold}
\title{Title}
\usage{
make_performance_metric_bold(hover_text, performance_metric)
}
\arguments{
\item{hover_text}{text}

\item{performance_metric}{perf}
}
\value{

}
\description{
Title
}
