% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision.R
\name{plot_decision_curve}
\alias{plot_decision_curve}
\title{Decision Curve from Performance Data}
\usage{
plot_decision_curve(
  performance_data,
  chosen_threshold = NA,
  interactive = TRUE,
  col_values = c("#1b9e77", "#d95f02", "#7570b3", "#e7298a", "#07004D", "#E6AB02",
    "#FE5F55", "#54494B", "#006E90", "#BC96E6", "#52050A", "#1F271B", "#BE7C4D",
    "#63768D", "#08A045", "#320A28", "#82FF9E", "#2176FF", "#D1603D", "#585123"),
  size = NULL,
  type = "conventional",
  min_p_threshold = 0,
  max_p_threshold = 1
)
}
\arguments{
\item{performance_data}{an rtichoke Performance Data}

\item{chosen_threshold}{a chosen threshold to display (for non-interactive)}

\item{interactive}{whether the plot should be interactive
plots}

\item{col_values}{color palette}

\item{size}{the size of the curve}

\item{type}{What type of Decision Curve, default choice is "conventional".
Alternatives are "interventions avoided" and "combined" for both
"conventional" and "interventions avoided" on the same view.}

\item{min_p_threshold}{The minimum Probability Threshold value to be
displayed}

\item{max_p_threshold}{The maximum Probability Threshold value to be
displayed}
}
\description{
Plot a Precision decision Curve
}
\examples{
\dontrun{


one_pop_one_model \%>\%
  plot_decision_curve()
  
one_pop_one_model \%>\%
  plot_decision_curve(type = "interventions avoided")
  
one_pop_one_model \%>\%
  plot_decision_curve(type = "combined")

multiple_models \%>\%
  plot_decision_curve()
  
multiple_models \%>\%
  plot_decision_curve(type = "interventions avoided")
  
multiple_models \%>\%
  plot_decision_curve(type = "combined")

multiple_populations \%>\%
  plot_decision_curve()
  
multiple_populations \%>\%
  plot_decision_curve(type = "interventions avoided")
  
multiple_populations \%>\%
  plot_decision_curve(type = "combined")
  
}
}
