% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision.R
\name{plot_decision_curve}
\alias{plot_decision_curve}
\title{Decision Curve from Performance Data}
\usage{
plot_decision_curve(
  performance_data,
  chosen_threshold = NA,
  interactive = F,
  main_slider = "threshold",
  col_values = c("#5BC0BE", "#FC8D62", "#8DA0CB", "#E78AC3", "#A4243B")
)
}
\arguments{
\item{performance_data}{an rtichoke Performance Data}

\item{chosen_threshold}{a chosen threshold to display (for non-interactive)}

\item{interactive}{whether the plot should be interactive
plots}

\item{main_slider}{what is the main slider - threshold, percent positives or positives}

\item{col_values}{color palette}
}
\description{
Plot a Precision decision Curve
}
\examples{

one_pop_one_model_as_a_vector \%>\%
  plot_decision_curve()

one_pop_one_model_as_a_vector_enforced_percentiles_symmetry \%>\%
  plot_decision_curve()

one_pop_one_model_as_a_list \%>\%
  plot_decision_curve()

one_pop_one_model_as_a_list_enforced_percentiles_symmetry \%>\%
  plot_decision_curve()

one_pop_three_models \%>\%
  plot_decision_curve()

one_pop_three_models_enforced_percentiles_symmetry \%>\%
  plot_decision_curve()

train_and_test_sets \%>\%
  plot_decision_curve()

train_and_test_sets_enforced_percentiles_symmetry \%>\%
  plot_decision_curve()
\dontrun{

one_pop_one_model_as_a_vector \%>\%
  plot_decision_curve(interactive = TRUE)

one_pop_one_model_as_a_vector_enforced_percentiles_symmetry \%>\%
  plot_decision_curve(interactive = TRUE, main_slider = "predicted_positives_percent")

one_pop_one_model_as_a_list \%>\%
  plot_decision_curve(interactive = TRUE)

one_pop_one_model_as_a_list_enforced_percentiles_symmetry \%>\%
  plot_decision_curve(interactive = TRUE, main_slider = "predicted_positives_percent")

one_pop_three_models \%>\%
  plot_decision_curve(interactive = TRUE)

one_pop_three_models_enforced_percentiles_symmetry \%>\%
  plot_decision_curve(interactive = TRUE, main_slider = "predicted_positives_percent")

train_and_test_sets \%>\%
  plot_decision_curve(interactive = TRUE)

train_and_test_sets_enforced_percentiles_symmetry \%>\%
  plot_decision_curve(interactive = TRUE, main_slider = "predicted_positives_percent")
}
}
