% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision.R
\name{plot_decision_curve}
\alias{plot_decision_curve}
\title{Decision Curve from Performance Data}
\usage{
plot_decision_curve(
  performance_data,
  chosen_threshold = NA,
  interactive = TRUE,
  col_values = c("#5BC0BE", "#FC8D62", "#8DA0CB", "#E78AC3", "#A4243B"),
  interventions_avoided = FALSE,
  size = NULL,
  type = "conventional"
)
}
\arguments{
\item{performance_data}{an rtichoke Performance Data}

\item{chosen_threshold}{a chosen threshold to display (for non-interactive)}

\item{interactive}{whether the plot should be interactive
plots}

\item{col_values}{color palette}

\item{interventions_avoided}{should one use interventions avoided curve}

\item{size}{the size of the curve}

\item{type}{What type of Decision Curve, default choice is "conventional".
Alternatives are "interventions avoided" and "combined" for both
"conventional" and "interventions avoided" on the same view.}
}
\description{
Plot a Precision decision Curve
}
\examples{
\dontrun{


one_pop_one_model \%>\%
  plot_decision_curve()

one_pop_one_model_by_ppcr \%>\%
  plot_decision_curve()

multiple_models \%>\%
  plot_decision_curve()

multiple_models_by_ppcr \%>\%
  plot_decision_curve()

multiple_populations \%>\%
  plot_decision_curve()

multiple_populations_by_ppcr \%>\%
  plot_decision_curve()
  
  
}
}
