% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gains.R
\name{plot_gains_curve}
\alias{plot_gains_curve}
\title{Gains Curve from Performance Data}
\usage{
plot_gains_curve(
  performance_data,
  chosen_threshold = NA,
  interactive = FALSE,
  main_slider = "threshold",
  col_values = c("#5BC0BE", "#FC8D62", "#8DA0CB", "#E78AC3", "#A4243B"),
  size = 350
)
}
\arguments{
\item{performance_data}{an rtichoke Performance Data}

\item{chosen_threshold}{a chosen threshold to display (for non-interactive)}

\item{interactive}{whether the plot should be interactive
plots}

\item{col_values}{color palette}

\item{size}{the size of the curve}
}
\description{
Plot a Gains Curve
}
\examples{

one_pop_one_model_as_a_vector \%>\%
  plot_gains_curve()

one_pop_one_model_as_a_vector_enforced_percentiles_symmetry \%>\%
  plot_gains_curve(main_slider = "ppcr")

one_pop_one_model_as_a_list \%>\%
  plot_gains_curve()

one_pop_one_model_as_a_list_enforced_percentiles_symmetry \%>\%
  plot_gains_curve(main_slider = "ppcr")

one_pop_three_models \%>\%
  plot_gains_curve()

one_pop_three_models_enforced_percentiles_symmetry \%>\%
  plot_gains_curve(main_slider = "ppcr")

train_and_test_sets \%>\%
  plot_gains_curve()

train_and_test_sets_enforced_percentiles_symmetry \%>\%
  plot_gains_curve(main_slider = "ppcr")
\dontrun{

one_pop_one_model_as_a_vector \%>\%
  plot_gains_curve(interactive = TRUE)

one_pop_one_model_as_a_vector_enforced_percentiles_symmetry \%>\%
  plot_gains_curve(interactive = TRUE, main_slider = "ppcr")

one_pop_one_model_as_a_list \%>\%
  plot_gains_curve(interactive = TRUE)

one_pop_one_model_as_a_list_enforced_percentiles_symmetry \%>\%
  plot_gains_curve(interactive = TRUE, main_slider = "ppcr")

one_pop_three_models \%>\%
  plot_gains_curve(interactive = TRUE)

one_pop_three_models_enforced_percentiles_symmetry \%>\%
  plot_gains_curve(interactive = TRUE, main_slider = "ppcr")

train_and_test_sets \%>\%
  plot_gains_curve(interactive = TRUE)

train_and_test_sets_enforced_percentiles_symmetry \%>\%
  plot_gains_curve(interactive = TRUE, main_slider = "ppcr")
}

}
