% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision.R
\name{plot_interventions_avoided}
\alias{plot_interventions_avoided}
\title{Decision Curve from Performance Data (Interventions Avoided)}
\usage{
plot_interventions_avoided(
  performance_data,
  chosen_threshold = NA,
  interactive = TRUE,
  col_values = c("#5BC0BE", "#FC8D62", "#8DA0CB", "#E78AC3", "#A4243B"),
  size = 350
)
}
\description{
Plot interventions Avoided
}
\examples{

\dontrun{

one_pop_one_model \%>\%
  plot_interventions_avoided()

one_pop_one_model_by_ppcr \%>\%
  plot_interventions_avoided()

multiple_models \%>\%
  plot_interventions_avoided()

multiple_models_by_ppcr \%>\%
  plot_interventions_avoided()

multiple_populations \%>\%
  plot_interventions_avoided()

multiple_populations_by_ppcr \%>\%
  plot_interventions_avoided()
  
}
}
\keyword{internal}
