% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision.R
\name{plot_interventions_avoided}
\alias{plot_interventions_avoided}
\title{Decision Curve from Performance Data (Interventions Avoided)}
\usage{
plot_interventions_avoided(
  performance_data,
  chosen_threshold = NA,
  interactive = TRUE,
  col_values = c("#5BC0BE", "#FC8D62", "#8DA0CB", "#E78AC3", "#A4243B"),
  size = 350
)
}
\description{
Plot interventions Avoided
}
\examples{

\dontrun{

one_pop_one_model_as_a_vector \%>\%
  plot_interventions_avoided()

one_pop_one_model_as_a_vector_enforced_percentiles_symmetry \%>\%
  plot_interventions_avoided(main_slider = "ppcr")

one_pop_one_model_as_a_list \%>\%
  plot_interventions_avoided()

one_pop_one_model_as_a_list_enforced_percentiles_symmetry \%>\%
  plot_interventions_avoided(main_slider = "ppcr")

one_pop_three_models \%>\%
  plot_interventions_avoided()

one_pop_three_models_enforced_percentiles_symmetry \%>\%
  plot_interventions_avoided(main_slider = "ppcr")

train_and_test_sets \%>\%
  plot_interventions_avoided()

train_and_test_sets_enforced_percentiles_symmetry \%>\%
  plot_interventions_avoided(main_slider = "ppcr")
}
}
\keyword{internal}
