% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lift.R
\name{plot_lift_curve}
\alias{plot_lift_curve}
\title{LIFT Curve from Performance Data}
\usage{
plot_lift_curve(
  performance_data,
  chosen_threshold = NA,
  interactive = TRUE,
  color_values = c("#1b9e77", "#d95f02", "#7570b3", "#e7298a", "#07004D", "#E6AB02",
    "#FE5F55", "#54494B", "#006E90", "#BC96E6", "#52050A", "#1F271B", "#BE7C4D",
    "#63768D", "#08A045", "#320A28", "#82FF9E", "#2176FF", "#D1603D", "#585123"),
  size = NULL
)
}
\arguments{
\item{performance_data}{an rtichoke Performance Data}

\item{chosen_threshold}{A numeric value for a specific threshold to highlight.}

\item{interactive}{A logical indicating whether to create an interactive \code{plotly} chart.}

\item{color_values}{A character vector of hex codes for the color palette.}

\item{size}{A numeric value for the line size.}
}
\description{
Plot a LIFT Curve
}
\examples{
\dontrun{

one_pop_one_model \%>\%
  plot_lift_curve()

one_pop_one_model_by_ppcr \%>\%
  plot_lift_curve()

multiple_models \%>\%
  plot_lift_curve()

multiple_models_by_ppcr \%>\%
  plot_lift_curve()

multiple_populations \%>\%
  plot_lift_curve()

multiple_populations_by_ppcr \%>\%
  plot_lift_curve()
}

}
