% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lift.R
\name{plot_lift_curve}
\alias{plot_lift_curve}
\title{LIFT Curve from Performance Data}
\usage{
plot_lift_curve(
  performance_data,
  chosen_threshold = NA,
  interactive = TRUE,
  col_values = c("#5BC0BE", "#FC8D62", "#8DA0CB", "#E78AC3", "#A4243B"),
  size = NULL
)
}
\arguments{
\item{performance_data}{an rtichoke Performance Data}

\item{chosen_threshold}{a chosen threshold to display (for non-interactive)}

\item{interactive}{whether the plot should be interactive
plots}

\item{col_values}{color palette}

\item{size}{the size of the curve}
}
\description{
Plot a LIFT Curve
}
\examples{

\dontrun{

one_pop_one_model \%>\%
  plot_lift_curve()

one_pop_one_model_by_ppcr \%>\%
  plot_lift_curve()

multiple_models \%>\%
  plot_lift_curve()

multiple_models_by_ppcr \%>\%
  plot_lift_curve()

multiple_populations \%>\%
  plot_lift_curve()

multiple_populations_by_ppcr \%>\%
  plot_lift_curve()
}

}
