% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_for_performance_metrics.R
\name{plot_precision_recall_curve}
\alias{plot_precision_recall_curve}
\title{Precision Recall Curve from Performance Table}
\usage{
plot_precision_recall_curve(
  performance_table,
  chosen_threshold = NA,
  interactive = F,
  main_slider = "threshold"
)
}
\arguments{
\item{performance_table}{an rtichoke performance table}

\item{chosen_threshold}{a chosen threshold to display}

\item{interactive}{whether the plot should be interactive}

\item{main_slider}{what is the main slider - threshold, percent positives or positives}
}
\description{
Plot a Precision Recall Curve
}
\examples{

performance_table_one_model \%>\%
  plot_precision_recall_curve()

performance_table_for_two_models \%>\%
  plot_precision_recall_curve()
  
performance_table_for_two_models \%>\%
  plot_precision_recall_curve()
  
}
