% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_curve.R
\name{plot_roc_curve}
\alias{plot_roc_curve}
\title{ROC Curve from Performance Data}
\usage{
plot_roc_curve(
  performance_data,
  chosen_threshold = NA,
  interactive = F,
  main_slider = "threshold",
  col_values = c("#21DACD", "#B6C174", "#A7DA2E", "#C2C172", "#FFD700")
)
}
\arguments{
\item{performance_data}{an rtichoke Performance Data}

\item{chosen_threshold}{a chosen threshold to display}

\item{interactive}{whether the plot should be interactive}

\item{main_slider}{what is the main slider - threshold, percent positives or positives}

\item{col_values}{color palette}
}
\description{
Plot a ROC Curve
}
\examples{

one_pop_one_model_as_a_vector \%>\%
  plot_roc_curve()

one_pop_one_model_as_a_vector_enforced_percentiles_symmetry \%>\%
  plot_roc_curve()

one_pop_one_model_as_a_list \%>\%
  plot_roc_curve()

one_pop_one_model_as_a_list_enforced_percentiles_symmetry \%>\%
  plot_roc_curve()

one_pop_three_models \%>\%
  plot_roc_curve()

one_pop_three_models_enforced_percentiles_symmetry \%>\%
  plot_roc_curve()

train_and_test_sets \%>\%
  plot_roc_curve()

train_and_test_sets_enforced_percentiles_symmetry \%>\%
  plot_roc_curve()
\dontrun{

one_pop_one_model_as_a_vector \%>\%
  plot_roc_curve(interactive = T)

one_pop_one_model_as_a_vector_enforced_percentiles_symmetry \%>\%
  plot_roc_curve(interactive = T, main_slider = "predicted_positives_percent")

one_pop_one_model_as_a_list \%>\%
  plot_roc_curve(interactive = T)

one_pop_one_model_as_a_list_enforced_percentiles_symmetry \%>\%
  plot_roc_curve(interactive = T, main_slider = "predicted_positives_percent")

one_pop_three_models \%>\%
  plot_roc_curve(interactive = T)

one_pop_three_models_enforced_percentiles_symmetry \%>\%
  plot_roc_curve(interactive = T, main_slider = "predicted_positives_percent")

train_and_test_sets \%>\%
  plot_roc_curve(interactive = T)

train_and_test_sets_enforced_percentiles_symmetry \%>\%
  plot_roc_curve(interactive = T, main_slider = "predicted_positives_percent")
}

}
