% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{plot_roc_curve}
\alias{plot_roc_curve}
\title{ROC Curve from Performance Data}
\usage{
plot_roc_curve(
  performance_data,
  chosen_threshold = NA,
  interactive = FALSE,
  main_slider = "threshold",
  col_values = c("#5BC0BE", "#FC8D62", "#8DA0CB", "#E78AC3", "#A4243B"),
  title_included = FALSE,
  size = NULL
)
}
\arguments{
\item{performance_data}{an rtichoke Performance Data}

\item{chosen_threshold}{a chosen threshold to display (for non-interactive)}

\item{interactive}{whether the plot should be interactive
plots}

\item{main_slider}{what is the main slider - threshold, percent positives
or positives}

\item{col_values}{color palette}

\item{title_included}{add title to the curve}

\item{size}{the size of the curve}
}
\description{
Plot a ROC Curve
}
\examples{

one_pop_one_model_as_a_vector \%>\%
  plot_roc_curve()

one_pop_one_model_as_a_vector_enforced_percentiles_symmetry \%>\%
  plot_roc_curve()

one_pop_one_model_as_a_list \%>\%
  plot_roc_curve()

one_pop_one_model_as_a_list_enforced_percentiles_symmetry \%>\%
  plot_roc_curve()

one_pop_three_models \%>\%
  plot_roc_curve()

one_pop_three_models_enforced_percentiles_symmetry \%>\%
  plot_roc_curve()

train_and_test_sets \%>\%
  plot_roc_curve()

train_and_test_sets_enforced_percentiles_symmetry \%>\%
  plot_roc_curve()
\dontrun{

one_pop_one_model_as_a_vector \%>\%
  plot_roc_curve(interactive = TRUE)

one_pop_one_model_as_a_vector_enforced_percentiles_symmetry \%>\%
  plot_roc_curve(interactive = TRUE, main_slider = "ppcr")

one_pop_one_model_as_a_list \%>\%
  plot_roc_curve(interactive = TRUE)

one_pop_one_model_as_a_list_enforced_percentiles_symmetry \%>\%
  plot_roc_curve(interactive = TRUE, main_slider = "ppcr")

one_pop_three_models \%>\%
  plot_roc_curve(interactive = TRUE)

one_pop_three_models_enforced_percentiles_symmetry \%>\%
  plot_roc_curve(interactive = TRUE, main_slider = "ppcr")

train_and_test_sets \%>\%
  plot_roc_curve(interactive = TRUE)

train_and_test_sets_enforced_percentiles_symmetry \%>\%
  plot_roc_curve(interactive = TRUE, main_slider = "ppcr")
}

}
