% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probs_hist.R
\name{prepare_probs_distribution_data}
\alias{prepare_probs_distribution_data}
\title{Prepare events per strata data}
\usage{
prepare_probs_distribution_data(
  probs,
  reals = NA,
  stratified_by = "probability_threshold",
  by = 0.01,
  condition_on = NA
)
}
\arguments{
\item{probs}{a list of vectors of estimated probabilities
(one for each model or one for each population)}

\item{reals}{a list of vectors of binary outcomes (one for each population)}

\item{stratified_by}{Performance Metrics can be stratified by Probability
Threshold or alternatively by Predicted Positives Condition Rate}

\item{by}{number: increment of the sequence.}
}
\description{
Prepare events per strata data
}
\examples{
\dontrun{

prepare_probs_distribution_data(
  probs = list(example_dat$estimated_probabilities),
  reals = list(example_dat$outcome)
)

# Several Models

prepare_probs_distribution_data(
  probs = list(
    "First Model" = example_dat$estimated_probabilities,
    "Second Model" = example_dat$random_guess
  ),
  reals = list(example_dat$outcome)
)


# Several Populations


prepare_probs_distribution_data(
  probs = list(
    "train" = example_dat \%>\%
      dplyr::filter(type_of_set == "train") \%>\%
      dplyr::pull(estimated_probabilities),
    "test" = example_dat \%>\% dplyr::filter(type_of_set == "test") \%>\%
      dplyr::pull(estimated_probabilities)
  ),
  reals = list(
    "train" = example_dat \%>\% dplyr::filter(type_of_set == "train") \%>\%
      dplyr::pull(outcome),
    "test" = example_dat \%>\% dplyr::filter(type_of_set == "test") \%>\%
      dplyr::pull(outcome)
  )
)
}
}
\keyword{internal}
