% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_for_performance_metrics.R
\name{remove_grid_lines_from_plotly}
\alias{remove_grid_lines_from_plotly}
\title{remove_grid_lines_from_plotly}
\usage{
remove_grid_lines_from_plotly(plotly_object)
}
\arguments{
\item{plotly_object}{a plotly plot for performance metrics}
}
\description{
remove_grid_lines_from_plotly
}
