% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_table.R
\name{render_and_format_gt}
\alias{render_and_format_gt}
\title{Rendering and Formatting gt}
\usage{
render_and_format_gt(
  performance_data,
  main_slider,
  perf_dat_type,
  prevalence,
  col_values
)
}
\arguments{
\item{performance_data}{an rtichoke Performance Data}

\item{main_slider}{what is the main slider - threshold, percent positives or positives}

\item{prevalence}{the prevalence of the populations}

\item{col_values}{color palette}
}
\value{

}
\description{
Rendering and Formatting gt
}
\keyword{internal}
