% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_table.R
\name{render_and_format_gt}
\alias{render_and_format_gt}
\title{Rendering and Formatting gt}
\usage{
render_and_format_gt(
  performance_data,
  main_slider,
  perf_dat_type,
  prevalence,
  color_values
)
}
\arguments{
\item{performance_data}{an rtichoke Performance Data}

\item{prevalence}{the prevalence of the populations}

\item{color_values}{color palette}
}
\description{
Rendering and Formatting gt
}
\keyword{internal}
