% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_table.R
\name{render_performance_table}
\alias{render_performance_table}
\title{Performance Table}
\usage{
render_performance_table(
  performance_data,
  chosen_threshold = NA,
  interactive = F,
  main_slider = "threshold",
  col_values = c("#5BC0BE", "#FC8D62", "#8DA0CB", "#E78AC3", "#A4243B")
)
}
\arguments{
\item{performance_data}{an rtichoke Performance Data}

\item{chosen_threshold}{a chosen threshold to display}

\item{interactive}{whether the plot should be interactive}

\item{main_slider}{what is the main slider - threshold, percent positives or positives}

\item{col_values}{color palette}
}
\description{
Create a Performance Table
}
\examples{

one_pop_one_model_as_a_vector \%>\%
  render_performance_table()

one_pop_one_model_as_a_vector_enforced_percentiles_symmetry \%>\%
  render_performance_table()

one_pop_one_model_as_a_list \%>\%
  render_performance_table()

one_pop_one_model_as_a_list_enforced_percentiles_symmetry \%>\%
  render_performance_table()

one_pop_three_models \%>\%
  render_performance_table()

one_pop_three_models_enforced_percentiles_symmetry \%>\%
  render_performance_table()

train_and_test_sets \%>\%
  render_performance_table()

train_and_test_sets_enforced_percentiles_symmetry \%>\%
  render_performance_table()
\dontrun{

one_pop_one_model_as_a_vector \%>\%
  render_performance_table(interactive = T)

one_pop_one_model_as_a_vector_enforced_percentiles_symmetry \%>\%
  render_performance_table(interactive = T)

one_pop_one_model_as_a_list \%>\%
  render_performance_table(interactive = T)

one_pop_one_model_as_a_list_enforced_percentiles_symmetry \%>\%
  render_performance_table(interactive = T)

one_pop_three_models \%>\%
  render_performance_table(interactive = T)

one_pop_three_models_enforced_percentiles_symmetry \%>\%
  render_performance_table(interactive = T)

train_and_test_sets \%>\%
  render_performance_table(interactive = T)

train_and_test_sets_enforced_percentiles_symmetry \%>\%
  render_performance_table(interactive = T)
}

}
