% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_table.R
\name{render_performance_table}
\alias{render_performance_table}
\title{Performance Table}
\usage{
render_performance_table(
  performance_data,
  chosen_threshold = NA,
  output_type = "reactable",
  col_values = c("#5BC0BE", "#FC8D62", "#8DA0CB", "#E78AC3", "#A4243B")
)
}
\arguments{
\item{performance_data}{an rtichoke Performance Data}

\item{chosen_threshold}{a chosen threshold to display (for non-interactive)}

\item{output_type}{the type of the output table}

\item{col_values}{color palette}
}
\description{
Create a Performance Table
}
\examples{
\dontrun{

one_pop_one_model \%>\%
  render_performance_table()

one_pop_one_model_by_ppcr \%>\%
  render_performance_table()

multiple_models \%>\%
  render_performance_table()

multiple_models_by_ppcr \%>\%
  render_performance_table()

multiple_populations \%>\%
  render_performance_table()

multiple_populations_by_ppcr \%>\%
  render_performance_table()

}

}
