% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtichoke-package.R
\docType{package}
\name{rtichoke-package}
\alias{rtichoke}
\alias{rtichoke-package}
\title{rtichoke: Interactive Visualization of Binary Prediction Model Performance}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools for exploring the performance of prediction models with binary outcomes through interactive and static visualizations. Supports receiver operating characteristic, precision-recall, gains, lift, decision, and calibration curves, together with performance tables and summary reports.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://uriahf.github.io/rtichoke/}
  \item \url{https://github.com/uriahf/rtichoke}
  \item Report bugs at \url{https://github.com/uriahf/rtichoke/issues}
}

}
\author{
\strong{Maintainer}: Uriah Finkel \email{ufinkel@gmail.com}

Authors:
\itemize{
  \item Uriah Finkel \email{ufinkel@gmail.com}
}

}
\keyword{internal}
