% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtichoke-package.R
\docType{package}
\name{rtichoke-package}
\alias{rtichoke}
\alias{rtichoke-package}
\title{rtichoke: Interactive Evaluation of Binary Prediction Models}
\description{
Interactive tools for evaluating binary prediction-model performance. Provides
visualizations and summaries for receiver operating characteristic and
precision-recall curves, gains and lift charts, decision curves, calibration,
confusion matrices, model metrics, and related performance reports.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://uriahf.github.io/rtichoke/}
  \item \url{https://github.com/uriahf/rtichoke}
  \item Report bugs at \url{https://github.com/uriahf/rtichoke/issues}
}

}
\author{
\strong{Maintainer}: Uriah Finkel \email{ufinkel@gmail.com}

Authors:
\itemize{
  \item Uriah Finkel \email{ufinkel@gmail.com}
}

}
\keyword{internal}
