% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_helper_functions.R
\name{set_axis_titles}
\alias{set_axis_titles}
\title{Set Titles for x and y axis in plotly objects}
\usage{
set_axis_titles(plotly_object, curve)
}
\arguments{
\item{plotly_object}{}

\item{curve}{}
}
\value{

}
\description{
Set Titles for x and y axis in plotly objects
}
