% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_helper_functions.R
\name{set_styling_for_rtichoke}
\alias{set_styling_for_rtichoke}
\title{Set styling for rtichoke plotly}
\usage{
set_styling_for_rtichoke(plotly_object, curve)
}
\arguments{
\item{plotly_object}{}

\item{curve}{}
}
\value{

}
\description{
Set styling for rtichoke plotly
}
